Time: 189 ms (8.06%) | Memory: 58.3 MB (97.63%) - LeetSync

+++ 2616-maximal-score-after-applying-k-operations/maximal-score-after-applying-k-operations.java
@@ -1,15 +1,15 @@
 class Solution {
     public long maxKelements(int[] nums, int k) {
-        PriorityQueue<Long> pq= new PriorityQueue<Long>(Collections.reverseOrder());
-        for(int i:nums){
-            pq.add((long)i);
+        PriorityQueue<Double> pq= new PriorityQueue<Double>(Collections.reverseOrder());
+        for(double i:nums){
+            pq.add(i);
         }
-        long result=0;
+        double result=0;
         for(int i=0;i<k;i++){
             double temp=pq.poll();
             result+=temp;
-            pq.add((long)Math.ceil(temp/3));
+            pq.add(Math.ceil(temp/3));
         }
-        return result;
+        return (long)result;
     }
 }
